Restructure, start writing tests. Return a png
