+++ src/lib.rs
@@ -1,41 +1,19 @@
 
 #![cfg_attr(feature="clippy", feature(plugin))]
 #![cfg_attr(feature="clippy", plugin(clippy))]
+#![feature(plugin, decl_macro)]
+#![plugin(rocket_codegen)]
 
-#[macro_use] extern crate diesel;
-#[macro_use] extern crate diesel_codegen;
 extern crate dotenv;
+#[macro_use] extern crate diesel;
+#[macro_use]extern crate diesel_codegen;
+extern crate rocket;
+extern crate rand;
+
 
 pub mod schema;
 pub mod models;
+pub mod images;
+pub mod server;
+pub mod db;
 
-use diesel::prelude::*;
-use diesel::pg::PgConnection;
-use dotenv::dotenv;
-
-use std::env;
-use self::models::{Post, NewPost};
-
-pub fn establish_connection() -> PgConnection {
-    dotenv().ok();
-
-    let database_url = env::var("DATABASE_URL")
-        .expect("DATABASE_URL must be set");
-
-    PgConnection::establish(&database_url)
-        .expect(&format!("Error connecting to {}", database_url))
-
-}
-
-pub fn create_post(conn: &PgConnection, title: &str, body: &str) -> Post {
-    use schema::posts;
-
-    let new_post = NewPost {
-        title,
-        body,
-    };
-
-    diesel::insert(&new_post).into(posts::table)
-        .get_result(conn)
-        .expect("Error saving new post")
-}
