+++ tests/integration_test.rs
@@ -1,0 +1,3 @@
+
+extern crate rust_360;
+
